bin/repl: run NodeJS REPL with initialised container
Example use:

    ./lib/bin/repl
    Available vars: Actees, Actors, Analytics, Assignments, Audits, Auth, Blobs, ClientAudits, Comments, Configs, Datasets, Entities, FieldKeys, FormAttachments, Forms, Keys, Projects, PublicLinks, Roles, Sentry, Sessions, SubmissionAttachments, Submissions, Users, all, db, enketo, env, mail, maybeOne, maybeOneFirst, one, oneFirst, run, s3, sql, stream, with, xlsform
    > await db.any(sql`SELECT * FROM blobs`)
    []
