+++ lib/bin/run-server.js
@@ -10,44 +10,16 @@
 // This is the main entrypoint for the actual HTTP server. It sets up the
 // dependency container and feeds it to the service infrastructure.
 
-const { mergeRight } = require('ramda');
 const config = require('config');
 const exit = require('express-graceful-exit');
 
 global.tap = (x) => { console.log(x); return x; }; // eslint-disable-line no-console
 
 ////////////////////////////////////////////////////////////////////////////////
-// CONTAINER SETUP
-
-// initialize our slonik connection pool.
-const { slonikPool } = require('../external/slonik');
-const db = slonikPool(config.get('default.database'));
-
-// set up our mailer.
-const env = config.get('default.env');
-const { mailer } = require('../external/mail');
-const mail = mailer(mergeRight(config.get('default.email'), { env }));
-
-// get a sentry and configure errors.
-const Sentry = require('../external/sentry').init(config.get('default.external.sentry'));
-Error.stackTrackLimit = 20;
-
-// get an xlsform client and a password module.
-const xlsform = require('../external/xlsform').init(config.get('default.xlsform'));
-
-// get an Enketo client
-const enketo = require('../external/enketo').init(config.get('default.enketo'));
-
-// get an S3 client.
-const s3 = require('../external/s3').init(config.get('default.external.s3blobStore'));
-
-
-////////////////////////////////////////////////////////////////////////////////
 // START HTTP SERVICE
 
 // initialize our container, then generate an http service out of it.
-const container = require('../model/container')
-  .withDefaults({ db, mail, env, Sentry, xlsform, enketo, s3 });
+const container = require('../util/default-container');
 const service = require('../http/service')(container);
 
 // insert the graceful exit middleware.
@@ -81,7 +53,7 @@ const term = () => {
   exit.gracefulExitHandler(service, server, {
     log: true,
     exitProcess: false,
-    callback: () => db.end().then(() => process.exit(0))
+    callback: () => container.db.end().then(() => process.exit(0))
   });
 };
